# scheduling-worker/documentation/openapi.yaml
components:
  schemas:
...
    AddMilestoneDto:
      properties: {}
      type: object
...
    CompleteMilestoneDto:
      properties: {}
      type: object
...
    CreateGoalDto:
      properties: {}
      type: object
...
    Goal:
      properties: {}
      type: object
    GoalProgressResponse:
      properties: {}
      type: object
...
    LogTimeDto:
      properties: {}
      type: object
...
    UpdateGoalDto:
      properties: {}
      type: object
...
    UpdateProgressDto:
      properties: {}
      type: object
  securitySchemes:
    jwt:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact: {}
  description: Goal management, habit tracking, calendar integration, and ML-powered scheduling
  title: Scheduling Worker API
  version: 1.0.0
openapi: 3.0.0
paths:
...
  /goals:
    get:
      operationId: GoalController_getGoals
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: status
          required: false
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Goal'
                type: array
          description: List of goals
      summary: Get all goals for a user
      tags:
        - Goals
    post:
      operationId: GoalController_createGoal
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalDto'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
          description: Goal created successfully
        '400':
          description: Invalid input
      summary: Create a new goal
      tags:
        - Goals
  /goals/user/{userId}/by-priority:
    get:
      operationId: GoalController_getGoalsByPriority
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Goal'
                type: array
          description: List of goals sorted by priority (high to low)
      summary: Get goals sorted by priority
      tags:
        - Goals
  /goals/user/{userId}/due-soon:
    get:
      operationId: GoalController_getGoalsDueSoon
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - description: 'Number of days ahead to look (default: 7)'
          in: query
          name: days
          required: false
          schema:
            type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Goal'
                type: array
          description: List of goals due soon
      summary: Get goals due soon for a user
      tags:
        - Goals
  /goals/user/{userId}/overdue:
    get:
      operationId: GoalController_getOverdueGoals
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Goal'
                type: array
          description: List of overdue goals
      summary: Get overdue goals for a user
      tags:
        - Goals
  /goals/{goalId}/milestones/{milestoneId}/complete:
    post:
      operationId: GoalController_completeMilestone
      parameters:
        - in: path
          name: goalId
          required: true
          schema:
            type: string
        - in: path
          name: milestoneId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteMilestoneDto'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
          description: Milestone completed successfully
        '404':
          description: Goal or milestone not found
      summary: Complete a milestone
      tags:
        - Goals
  /goals/{id}:
    delete:
      operationId: GoalController_deleteGoal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Goal deleted successfully
        '404':
          description: Goal not found
      summary: Delete a goal
      tags:
        - Goals
    get:
      operationId: GoalController_getGoal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
          description: Goal details
        '404':
          description: Goal not found
      summary: Get a specific goal by ID
      tags:
        - Goals
    put:
      operationId: GoalController_updateGoal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGoalDto'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
          description: Goal updated successfully
        '404':
          description: Goal not found
      summary: Update a goal
      tags:
        - Goals
  /goals/{id}/milestones:
    post:
      operationId: GoalController_addMilestone
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMilestoneDto'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
          description: Milestone added successfully
        '404':
          description: Goal not found
      summary: Add a milestone to a goal
      tags:
        - Goals
  /goals/{id}/progress:
    post:
      operationId: GoalController_updateProgress
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProgressDto'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
          description: Progress updated successfully
        '404':
          description: Goal not found
      summary: Update goal progress
      tags:
        - Goals
  /goals/{id}/progress-summary:
    get:
      operationId: GoalController_getProgressSummary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GoalProgressResponse'
          description: Goal progress summary
        '404':
          description: Goal not found
      summary: Get progress summary for a goal
      tags:
        - Goals
  /goals/{id}/time:
    post:
      operationId: GoalController_logTime
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogTimeDto'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
          description: Time logged successfully
        '404':
          description: Goal not found
      summary: Log time spent on a goal
      tags:
        - Goals
...
servers: []
...